Robust solution for checking the device SDK in ios_test_runner
Robust solution for checking the device SDK in ios_test_runner.
- Support passing optional argument --platform=[ios_device,ios_simulator].
- If args.platform is not given, use the output `xcrun simctl list devices` and `instruments -s devices` to detect the SDK of the device from given device id.

For new iOS models, such as iPhone Xs, Xs MAX, the uuid is no longer 40 digit. To be more robust, ios_test_runner won't use uuid format to determine the SDK of the device. To be consistent with Google bazel, we add a new argument --platform.