Add platform's usr/lib directory to UI tests
With Xcode 11 there's a new module called libXCTestSwiftSupport.dylib
that is required for running UI tests with Swift. This library is passed
to the Swift compiler / the linker with
https://github.com/bazelbuild/rules_swift/pull/300 but it also needs to
be accessible at runtime. Xcode copies this into the runner app's
Frameworks directory, but this is good enough and a bit simpler.

This might also be needed for some cases of logic / unit tests