Fix python2 interpreter
By default when specifying PY2 as the `python_version` the shebang in
the produced parfile is `/usr/bin/env python2` which doesn't exist on
macOS (unless users have created it).

This fix forces the shebang to be `/usr/bin/python2.7` which exists on
all recent macOS versions. Once this project is converted to python3,
we should change this to `/usr/bin/python3` instead.