Stop including stderr in simulator output
On M1 macs running anything with xcrun prints a bunch of these warnings
to stderr (FB9089778):

```
objc[35011]: Class AMSupportURLConnectionDelegate is implemented in both /usr/lib/libauthinstall.dylib (0x1ffc8ab90) and /Library/Apple/System/Library/PrivateFrameworks/MobileDevice.framework/Versions/A/MobileDevice (0x1044502c8). One of the two will be used. Which one is undefined.
```

So capturing this output, and then using it as json, is invalid on M1
machines when this happens. None of the consumers of this function check
the output for any error strings, so this change should only improve the
other use cases.