Add ability to use `test --id` param for sim name
As the title states, add the ability to use the `--id` param in the `test` command for the simulator name (new) or identifier (existing). If the param is a valid UUID, this change assumes the value is an identifier. There’s room to either add a proper `--name` arg in the future but this seemed the easiest to allow pass-through from places that already forward identifier like `rules_apple` and so on.