feat: Retry fetching simulator OS versions to fix flaky failures
Fixes flaky failures with the error:
`IllegalArgumentError: ... supported simulator os versions are [].`

This appears to be a race condition where an early call to `simctl`
returns an empty list of runtimes. Subsequent calls succeed.

This change makes the `GetSupportedSimOsVersions` function more
robust by retrying the call up to 5 times with a 1-second delay
between each attempt.